todo1 feito, logica de rotas dependendo do endpoint de login
