+++ project-frontend-ilab-eqp3/package.json
@@ -10,7 +10,9 @@
   "dependencies": {
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-toastify": "^8.2.0"
+    "react-router-dom": "^5.3.0",
+    "react-toastify": "^8.2.0",
+    "react-use": "^17.3.2"
   },
   "devDependencies": {
     "@vitejs/plugin-react": "^1.0.7",
