+++ project-frontend-ilab-eqp3/src/App.jsx
@@ -1,95 +1,45 @@
-import Button from './components/Button'
-import {useState} from 'react';
-import 'react-toastify/dist/ReactToastify.css';
-import { ToastContainer, toast } from 'react-toastify';
+import React from "react";
+import Login from './pages/Login';
+import Home from './pages/Home';
+import useUser from './hooks/useUser';
+
+
+import {
+  BrowserRouter,
+  Route,
+  Switch,
+  Redirect
+} from 'react-router-dom'
+
 
 
 function App() {
+  //const { token } = useUser();
 
-  const [email, setEmail] = useState('');
-  const [senha, setSenha] = useState('');
-  const [usuarioNaoExiste, setUsuarioNaoExiste] = useState(true);
+  //TODO: Apos ter a logica de login, refatorar para puxar o token e armazenar
 
-  async function handleLogin(e) {
-    e.preventDefault();
+  const token = false;
 
-    if(usuarioNaoExiste) {
-      return toast.error('Usuário não existe!', {
-        position: "top-right",
-        autoClose: 5000,
-        hideProgressBar: false,
-        closeOnClick: true,
-        pauseOnHover: false,
-        draggable: true,
-        progress: undefined,
-      });
-    }
-
-    return toast.success('Login bem sucedido!', {
-      position: "top-right",
-      autoClose: 5000,
-      hideProgressBar: false,
-      closeOnClick: true,
-      pauseOnHover: false,
-      draggable: true,
-      progress: undefined,
-    });
+  function RotasProtegidas(props) {
+    return (
+      <Route
+        render={() => token ? (props.children) : (<Redirect to="/" />)}
+      />
+    )
   }
 
-
   return (
-    <main className="tela_login">
-      <ToastContainer
-        className="toast-error"
-        position="top-right"
-        autoClose={2000}
-        hideProgressBar={false}
-        newestOnTop={false}
-        closeOnClick
-        rtl={false}
-        pauseOnFocusLoss
-        draggable
-        pauseOnHover={false}
-        theme="colored"
-      />
+    <BrowserRouter>
+      <Switch>
+        <Route path="/login" component={token ? Home : Login} />
+        <Route path="/" exact component={token ? Home : Login} />
+        <RotasProtegidas>
+          <Route path="/home" component={Home} />
+        </RotasProtegidas>
 
-      <section className="tela_login_direita">
-        <h2 className="tela_login_direita_titulo">Faça seu login!</h2>
-
-        <form onSubmit={handleLogin}>
-          <div className="form_input email_login">
-            <label htmlFor="input-email" className="form_label_login">E-mail</label>
-            <input
-              className="input_login"
-              type="email"
-              name="input-email"
-              placeholder="Digite seu e-mail"
-              value={email}
-              onChange={e => setEmail(e.target.value)}
-              required />
-          </div>
-
-          <div className="form_input senha_login botao">
-            <label htmlFor="input-senha" className="form_label_login">Senha</label>
-            <input
-              className="input_login"
-              type="password"
-              name="input-senha"
-              placeholder="Digite sua senha"
-              value={senha}
-              onChange={e => setSenha(e.target.value)}
-              required
-            />
-          </div>
-
-          <div className="tela_login_direita_div_botao" >
-              <Button texto={"Entrar"} />
-          </div>
-
-        </form>
-      </section>
-    </main>
-  )
+      </Switch>
+    </BrowserRouter>
+  );
 }
 
-export default App
+export default App;
