+++ project-frontend-ilab-eqp3/src/hooks/useUser.jsx
@@ -1,0 +1,8 @@
+import { useContext } from 'react';
+import UserContext from '../context/UserContext';
+
+function useUser() {
+  return useContext(UserContext);
+}
+
+export default useUser;
