+++ project-frontend-ilab-eqp3/src/index.css
@@ -25,64 +25,3 @@ code {
 
 
 /**/
-
-.tela_login {
-  height: 100vh;
-  width: 100vw;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background-color: #F8F8F9;
-}
-
-
-.tela_login_direita_titulo {
-  font-family: Montserrat;
-  font-style: normal;
-  font-weight: bold;
-  font-size: 2.4rem;
-  line-height: 130%;
-  color: #343447;
-  margin-bottom: 5.8rem;
-}
-
-form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-input {
-  width: 344px;
-  font-size: 1.6rem;
-}
-
-.form_input {
-  display: flex;
-  flex-direction: column;
-}
-
-.input_login {
-  border: 1.5px solid #D0D5DD; 
-  box-shadow: 0px 1px 2px rgba(16, 24, 40, 0.05);
-  border-radius: 8px;
-  padding: 1rem 1.4rem;
-}
-
-.email_login {
-  margin-bottom: 2rem;
-}
-
-.form_label_login {
-  font-family: Nunito;
-  font-weight: 600;
-  font-size: 1.4rem;
-  line-height: 20px;
-  color: #344054;
-  margin-bottom: 0.6rem;
-}
-
-
-.tela_login_direita_div_botao {
-  margin-top: 4.6rem;
-}
