+++ project-frontend-ilab-eqp3/src/main.jsx
@@ -2,10 +2,13 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import './index.css'
 import App from './App'
+import {UserProvider} from './context/userContext';
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <UserProvider> 
+      <App />
+    </UserProvider>
   </React.StrictMode>,
   document.getElementById('root')
 )
