+++ project-frontend-ilab-eqp3/src/pages/Home/index.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Home() {
+  return (
+    <div>
+      <h1>Home</h1>
+    </div>
+  )
+}
